+++ src/sys/idt.rs
@@ -138,7 +138,7 @@ extern "x86-interrupt" fn page_fault_handler(
     };
 
     if error_code.contains(PageFaultErrorCode::CAUSED_BY_WRITE) {
-        if sys::allocator::alloc_pages(&mut mapper, addr, 1).is_err() {
+        if sys::mem::alloc_pages(&mut mapper, addr, 1).is_err() {
             printk!(
                 "{}Error:{} Could not allocate page at {:#X}\n",
                 csi_color, csi_reset, addr
@@ -154,7 +154,7 @@ extern "x86-interrupt" fn page_fault_handler(
         // longer a simple clone of the kernel page table. Currently a process
         // is executed from its kernel address that is shared with the process.
         let start = (addr / 4096) * 4096;
-        if sys::allocator::alloc_pages(&mut mapper, start, 4096).is_ok() {
+        if sys::mem::alloc_pages(&mut mapper, start, 4096).is_ok() {
             if sys::process::is_userspace(start) {
                 let code_addr = sys::process::code_addr();
                 let src = (code_addr + start) as *mut u8;
