+++ src/sys/net/nic/pcnet.rs
@@ -1,5 +1,5 @@
 use crate::sys;
-use crate::sys::allocator::PhysBuf;
+use crate::sys::mem::PhysBuf;
 use crate::sys::net::{Config, EthernetDeviceIO, Stats};
 
 use alloc::sync::Arc;
