+++ src/sys/net/nic/rtl8139.rs
@@ -1,4 +1,4 @@
-use crate::sys::allocator::PhysBuf;
+use crate::sys::mem::PhysBuf;
 use crate::sys::net::{Config, EthernetDeviceIO, Stats};
 
 use alloc::sync::Arc;
